requirements.txt
Django
djangorestframework==3.15.2
pillow